requirements.txt
streamlit>=1.20
openai>=1.0
docxtpl>=0.16.4
python-docx>=0.8.11
pandas>=2.0
pydantic>=1.10
requests>=2.28
python-dotenv>=1.0
reportlab>=4.4
truststore>=0.10
